feat: Add expiration time to auction when starting auction

+++ backend/AuctionApi/Models/Auction.cs
@@ -25,7 +25,7 @@ public class Auction
     public int? MinimumPrice { get; set; }
 
     [BsonElement("expiration")]
-    public int? Expiration { get; set; }
+    public long? Expiration { get; set; }
 
     [BsonElement("itemName")]
     public string? ItemName { get; set; }
